:art: Change sidebar letters styles
Rain effect is now removed because it caused too much performance issue and it was also temporary. Instead of this sidebar is now in global scope. It means you don't need to add it on every new section/component. Also, to recompense the rain effect, the sidebar has now scrolling effect and more various content.