:art: Improve initial animation and DOM rendering
Before was the DOM of the website content in page loader not available. It was added after the initial page loader was done. This is not optimal for search engines cause there was no content of the website in site source code. Now everything is rendered on initial page load. Also, the fluidity between initial Page load and staging animation was improved